Rename pkg/datachannel to pkg/dcep in favor of consistency
