+++ pkg/dcep/message_test.go
@@ -1,4 +1,4 @@
-package datachannel
+package dcep
 
 import (
 	"testing"
